Organize parameter API into Repl.Parameters namespace and folders
