+++ src/Repl.Core/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using Repl.Parameters;
