+++ src/Repl.Core/Parameters/Attributes/ReplArgumentAttribute.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Parameters;
 
 /// <summary>
 /// Configures positional argument metadata for a handler parameter.
