+++ src/Repl.Core/Parameters/Attributes/ReplValueAliasAttribute.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Parameters;
 
 /// <summary>
 /// Maps an alias token to an injected option value for a parameter.
