+++ src/Repl.IntegrationTests/GlobalUsings.cs
@@ -1,2 +1,3 @@
 global using AwesomeAssertions;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Repl.Parameters;
